+++ sys/behave/behave.cabal
@@ -17,7 +17,6 @@ library
 
   exposed-modules:      Behave
                       , Expand
-                      , Subst
 
   other-modules:        Equiv
                       , Next
@@ -54,7 +53,6 @@ test-suite behave-test
                       , Expand
                       , Next
                       , Reduce
-                      , Subst
                       , Unfold
                       , TestBExpr
                       , TestFuncContent
