+++ sys/bexpr/src/Subst.hs
@@ -24,16 +24,17 @@ import qualified Data.Map as Map
 import qualified Data.Set as Set
 
 -- TorXakis imports
+import           BehExprDefs
 import           FuncDef
 import           FuncId
-import           TxsDefs
 import qualified ValExpr
+import           VarEnv
 import           VarId
 
 -- | Expressions that support substitution of variables for expressions.
 class Subst e where
     -- | Substitution function.
-    subst :: TxsDefs.VEnv                   -- ^ Mapping from variable id's to
+    subst :: VEnv                   -- ^ Mapping from variable id's to
                                             -- expressions on those variable id's.
           -> Map.Map FuncId (FuncDef VarId) -- ^ Mapping of function identifiers.
                                             -- to their definitions.
@@ -48,9 +49,9 @@ instance (Ord e, Subst e) => Subst (Set.Set e) where
 
 instance Subst BExpr where
     subst ve _     | Map.null ve = id
-    subst ve fdefs               = subst' ve fdefs . TxsDefs.view
+    subst ve fdefs               = subst' ve fdefs . view
     
-subst' :: TxsDefs.VEnv -> Map.Map FuncId (FuncDef VarId) -> BExprView -> BExpr
+subst' :: VEnv -> Map.Map FuncId (FuncDef VarId) -> BExprView -> BExpr
 subst' ve fdefs (ActionPref (ActOffer offs hidvars cnrs) bexp) =
     actionPref (ActOffer (subst ve fdefs offs)
                          hidvars
